A Toolkit for Integrated Species Distribution Models in R

isdmtools provides a set of tools for preparing, analyzing and visualizing spatial data for integrated species distribution models (ISDMs). It been developed to assist users in preparing multisource spatial point datasets for spatial block cross-validation, analysing the habitat suitability from the joint model predictions, with a focus on Bayesian inference and visualising the results. The package also provides a holistic view of model performance by computing a comprehensive evaluation metrics, including ROC-based and continuous-outcome weighted composite scores.

See also

Author

Maintainer: Akoeugnigan Idelphonse SODE sdidelphonse@gmail.com (ORCID) (SODE A. Idelphonse - continued development of the main code)

Other contributors:

Adandé Belarmain Fandohan [contributor]

Elias Teixeira Krainski [contributor]

Romain Glèlè Kakaï [contributor]